+++ PiwikPlugin/PiwikPlugin.php
@@ -26,6 +26,10 @@ class PiwikPlugin extends PluginBase {
     static protected $name = 'Piwik for Limesurvey';
 
     protected $settings = array(
+        'piwik_title_BasicSettings'=>array(
+            'type'=>'info',
+            'content'=>'<legend><small>Basic Piwik settings</small></legend><p>Global settings for this plugin</p>'
+        ),
         'piwik_trackAdminPages' => array(
             'type'=>'select',
             'options'=>array(0=>'No',1=>'Yes'),
@@ -56,6 +60,32 @@ class PiwikPlugin extends PluginBase {
             'default'=> 1
         ),
 */
+
+        
+        //---------- Event and Content Tracking settings----------
+        'piwik_title_EventAndContentTracking'=>array(
+            'type'=>'info',
+            'content'=>'<legend><small>Content and Event tracking</small></legend>'
+        ),
+        'piwik_trackContent' => array(
+            'type'=>'select',
+            'options'=>array(0=>'No',1=>'Yes'),
+             'label'=>'Track respondents\' interactions with answer options<br/><small>Uses Piwik <i>content tracking</i> on answer options</small>',
+            'default'=> 1
+        ),        
+        'piwik_trackEvents' => array(
+            'type'=>'select',
+            'options'=>array(0=>'No',1=>'Yes'),
+             'label'=>'Track important survey events<br/><small>Uses Piwik <i>event tracking</i> on selected survey events</small>',
+            'default'=> 1
+        ),
+        'piwik_trackEventsCategory' => array(
+            'type'=>'string',
+             'label'=>'Piwik Event category to use for survey events',
+            'default'=> 'Survey'
+        ),
+        
+        
     );
 
     private $registeredTrackingCode;
@@ -101,6 +131,7 @@ class PiwikPlugin extends PluginBase {
         else
         {
             // Update piwik_CustomUrl script
+			$this->loadEventTracking_moveButtons(); //Track use of prev/back/save/clear buttons in a survey.
         }
     }
 
@@ -114,10 +145,16 @@ class PiwikPlugin extends PluginBase {
                 $this->settings['piwik_trackSurveyPages']['default'] // If global is not set get the 'default' setting
             )
         );
+        $eventTracking=$this->get('piwik_trackEvents',null,null,false);
         if($trackThisSurvey)
         {
-            $js="_paq.push(['trackEvent', 'Survey', 'Survey-$iSurveyId', 'Submitted']);"; //Allows comparison of submit rates
-            App()->getClientScript()->registerScript('piwikPlugin_Event_Completed',$js,CClientScript::POS_END);
+	        if ($eventTracking){
+		        $eventCategory=$this->get('piwik_trackEventsCategory',null,null,false);
+                $js="_paq.push(['trackEvent', '$eventCategory', 'Survey-$iSurveyId', 'Completed']);"; //Event tracking of completed surveys.
+				App()->getClientScript()->registerScript('piwikPlugin_Event_Completed',$js,CClientScript::POS_END);
+	        }
+	        
+	        
             $piwikCustomUrl="survey/{$iSurveyId}/completed";
             App()->getClientScript()->registerScript('piwikCustomUrl',"_paq.push(['setCustomUrl', '{$piwikCustomUrl}'])",CClientScript::POS_END);
         }
@@ -126,6 +163,7 @@ class PiwikPlugin extends PluginBase {
     function unloadPiwikTrackingCode(){
             App()->getClientScript()->registerScript('piwikPlugin_TrackingCode','',CClientScript::POS_END);
             $this->registeredTrackingCode=false; //false
+            //Todo: Unload event tracking, too. Low priority.
     }
 
 
@@ -207,7 +245,55 @@ class PiwikPlugin extends PluginBase {
 
         // Add a script piwikCustomUrl at begin of body. Just to set the piwikCustomUrl: this var can be updated in another function
         App()->getClientScript()->registerScript('piwikCustomUrl',"_paq.push(['setCustomUrl', '{$piwikCustomUrl}'])",CClientScript::POS_END);
+		$this->loadContentTracking_questionanswers();
     }
+
+	function loadEventTracking_moveButtons(){
+		//Adds event tracking code to the #moveprevbtn and #movenextbtn.  
+		//see https://developer.piwik.org/api-reference/events
+		//This assumes the IDs in application/helpers/frontend_helper.php will remain constant.
+		
+        if ($eventTracking){
+	        $eventCategory=$this->get('piwik_trackEventsCategory',null,null,false);
+            $js=
+"$('#clearall').on('click',function(){ _paq.push(['trackEvent', '$eventCategory', 'Survey-$iSurveyId', 'Button: Clear responses']); });
+$('#saveallbtn').on('click',function(){ _paq.push(['trackEvent', '$eventCategory', 'Survey-$iSurveyId', 'Button: Save responses']); });
+$('#moveprevbtn').on('click',function(){ _paq.push(['trackEvent', '$eventCategory', 'Survey-$iSurveyId', 'Button: Previous page']); });
+$('#movenextbtn').on('click',function(){_paq.push(['trackEvent', '$eventCategory', 'Survey-$iSurveyId', 'Button: Next page']);});"; //TOODO: add the page it was used on. 
+			//Do not add an event to'#movesubmitbtn': this is done in the afterSurveyComplete event once all required answers have been answered and submission is confirmed.
+			App()->getClientScript()->registerScript('piwikPlugin_Event_ButtonUse',$js,CClientScript::POS_END);
+		}
+	}
+
+	function loadContentTracking_questionanswers(){
+		//Tags question input elements as content, to track interactions with. 
+		//See http://developer.piwik.org/guides/content-tracking
+		$trackContent=$this->get('piwik_trackContent',null,null,false);
+		if ($trackContent){
+	        $iSurveyId=$this->getParam('sid');
+			if(!$iSurveyId)
+            	$iSurveyId=$this->getParam('surveyid');
+            	
+			//Tags all inputs as content, to track content interactions
+			$js="_paq.push(['trackPageView']);
+	_paq.push(['trackVisibleContentImpressions']);"; //enable tracking
+			$js.="
+	//Define the content blocks
+	$('.question-wrapper').each(function(){
+		$(this).attr('data-track-content',''); 
+		$(this).attr('data-content-name','Survey-$iSurveyId');
+	});
+	//Add the item pieces
+ 	$('.question-wrapper').find('input').each
+	(function()
+		{ 
+		itemName=$(this).attr('name').toLowerCase();
+		$(this).attr('data-content-piece',itemName);
+		}
+	);"; 		
+        	App()->getClientScript()->registerScript('piwikPlugin_InputContentTracking',$js,CClientScript::POS_END);			
+		}
+	}
 
     public function beforeSurveySettings()
     {
@@ -215,6 +301,10 @@ class PiwikPlugin extends PluginBase {
         $event->set("surveysettings.{$this->id}", array(
             'name' => get_class($this),
             'settings' => array(
+                'piwik_Title_TrackingSettings'=>array(
+                    'type'=>'info',
+                    'content'=>'<legend><small>Tracking settings</small></legend>'
+                ),
                 'piwik_trackThisSurvey' => array(
                     'type' => 'select',
                     'options'=>array(